internal/apiService:extract balanceService->internal/balanceService
Extracted the balance service from the api service.

These two services should have separate implementations.
Splitting them provides for better protability, logic seperation and
security as only the spi server's api port, and the slb listenport
should be exposed.

Both the apiService as well as the balanceService both have GrpcServers,
which receive the same api calls, but the api service shall handle
request validation etc.
