+++ internal/apiService/service_test.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"testing"
 
+	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
@@ -16,142 +17,115 @@ const (
 	serverPort   = ":8081"
 )
 
-func setupServer() (*grpc.Server, *balanceServer) {
+type mockClient struct {
+	gen.BalanceClient
+	config *gen.Config
+}
+
+func (c *mockClient) Configuration(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*gen.Config, error) {
+	return c.config, nil
+}
+
+func (c *mockClient) Configure(ctx context.Context, in *gen.Config, opts ...grpc.CallOption) (*empty.Empty, error) {
+	c.config = in
+	return nil, nil
+}
+
+func (c *mockClient) Run(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*empty.Empty, error) {
+	return nil, nil
+}
+
+func (c *mockClient) Stop(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*empty.Empty, error) {
+	return nil, nil
+}
+func (c *mockClient) Add(ctx context.Context, in *gen.Server, opts ...grpc.CallOption) (*empty.Empty, error) {
+	return nil, nil
+}
+func (c *mockClient) Remove(ctx context.Context, in *gen.Server, opts ...grpc.CallOption) (*empty.Empty, error) {
+	return nil, nil
+}
+
+func setupServer(client gen.BalanceClient) (*grpc.Server, *BalanceServer) {
 	grpcServer := grpc.NewServer()
-	balanceServer := &balanceServer{}
+	balanceServer := &BalanceServer{Client: client}
 	gen.RegisterBalanceServer(grpcServer, balanceServer)
 	return grpcServer, balanceServer
 }
 
-func TestConfigurationNoSLBShouldReturnError(t *testing.T) {
-	_, balanceServer := setupServer()
-
-	ctx := context.Background()
-	_, err := balanceServer.Configuration(ctx, &emptypb.Empty{})
-
-	require.Error(t, err)
-	require.Equal(t, ErrNotConfigured, err)
-}
-
-func TestConfigurationWithSLBShouldReturnConfig(t *testing.T) {
-	_, balanceServer := setupServer()
-	slbConfig := &gen.Config{
-		ListenAddress: localAddress,
-		ListenPort:    defaultPort,
-		Endpoints:     []*gen.Server{{Address: localAddress}},
-		Strategy:      gen.SelectorStrategy_SELECTOR_STRATEGY_RANDOM,
-	}
-	_, err := balanceServer.Configure(context.Background(), slbConfig)
-	require.NoError(t, err)
-	config, err := balanceServer.Configuration(context.Background(), &emptypb.Empty{})
-
-	require.NoError(t, err)
-	require.Equal(t, slbConfig.ListenAddress, config.ListenAddress)
-	require.Equal(t, slbConfig.ListenPort, config.ListenPort)
-	require.Exactly(t, gen.SelectorStrategy_SELECTOR_STRATEGY_RANDOM, config.Strategy)
-}
-
 func TestConfigureShouldSetNewConfig(t *testing.T) {
-	_, balanceServer := setupServer()
+	_, apiService := setupServer(&mockClient{})
 	newConfig := &gen.Config{
 		ListenAddress: localAddress,
 		ListenPort:    "9090",
 		Endpoints:     []*gen.Server{{Address: localAddress}},
 	}
 
-	_, err := balanceServer.Configure(context.Background(), newConfig)
+	_, err := apiService.Configure(context.Background(), newConfig)
 	require.NoError(t, err)
-}
-
-func TestRunNoSLBShouldReturnError(t *testing.T) {
-	_, balanceServer := setupServer()
-
-	ctx := context.Background()
-	_, err := balanceServer.Run(ctx, &emptypb.Empty{})
-
-	require.Error(t, err)
-	require.Equal(t, ErrNotConfigured, err)
 }
 
 func TestRunWithSLBShouldRunSuccessfully(t *testing.T) {
-	_, balanceServer := setupServer()
+	_, apiService := setupServer(&mockClient{})
 	slbConfig := &gen.Config{
 		ListenAddress: localAddress,
 		ListenPort:    defaultPort,
 		Endpoints:     []*gen.Server{{Address: localAddress}},
 	}
 
-	balanceServer.Configure(context.Background(), slbConfig)
-	_, err := balanceServer.Run(context.Background(), &emptypb.Empty{})
+	apiService.Configure(context.Background(), slbConfig)
+	_, err := apiService.Run(context.Background(), &emptypb.Empty{})
 	require.NoError(t, err)
 }
 
-func TestStopNoSLBShouldReturnError(t *testing.T) {
-	_, balanceServer := setupServer()
+func TestRunWithSLBShouldFailRunWithoutEndpoints(t *testing.T) {
+	_, apiService := setupServer(&mockClient{})
+	slbConfig := &gen.Config{
+		ListenAddress: localAddress,
+		ListenPort:    defaultPort,
+		Endpoints:     []*gen.Server{},
+	}
 
-	ctx := context.Background()
-	_, err := balanceServer.Stop(ctx, &emptypb.Empty{})
-
+	apiService.Configure(context.Background(), slbConfig)
+	_, err := apiService.Run(context.Background(), &emptypb.Empty{})
 	require.Error(t, err)
-	require.Equal(t, ErrNotConfigured, err)
 }
 
 func TestStopWithSLBShouldStopSuccessfully(t *testing.T) {
-	_, balanceServer := setupServer()
+	_, apiService := setupServer(&mockClient{})
 	slbConfig := &gen.Config{
 		ListenAddress: localAddress,
 		ListenPort:    defaultPort,
 		Endpoints:     []*gen.Server{{Address: localAddress}},
 	}
 
-	balanceServer.Configure(context.Background(), slbConfig)
-	_, err := balanceServer.Stop(context.Background(), &emptypb.Empty{})
+	apiService.Configure(context.Background(), slbConfig)
+	_, err := apiService.Stop(context.Background(), &emptypb.Empty{})
 	require.NoError(t, err)
 }
 
-func TestAddNoSelectorShouldReturnError(t *testing.T) {
-	_, balanceServer := setupServer()
-
-	ctx := context.Background()
-	_, err := balanceServer.Add(ctx, &gen.Server{Address: localAddress + serverPort})
-
-	require.Error(t, err)
-	require.Equal(t, ErrNotConfigured, err)
-}
-
 func TestAddWithSelectorShouldAddServer(t *testing.T) {
-	_, balanceServer := setupServer()
+	_, apiService := setupServer(&mockClient{})
 	slbConfig := &gen.Config{
 		ListenAddress: localAddress,
 		ListenPort:    defaultPort,
 		Strategy:      gen.SelectorStrategy_SELECTOR_STRATEGY_ROUND_ROBIN,
 	}
 
-	balanceServer.Configure(context.Background(), slbConfig)
-	_, err := balanceServer.Add(context.Background(), &gen.Server{Address: localAddress + serverPort})
+	apiService.Configure(context.Background(), slbConfig)
+	_, err := apiService.Add(context.Background(), &gen.Server{Address: localAddress + serverPort})
 	require.NoError(t, err)
 }
 
-func TestRemoveNoSelectorShouldReturnError(t *testing.T) {
-	_, balanceServer := setupServer()
-
-	ctx := context.Background()
-	_, err := balanceServer.Remove(ctx, &gen.Server{Address: localAddress + serverPort})
-
-	require.Error(t, err)
-	require.Equal(t, ErrNotConfigured, err)
-}
-
 func TestRemoveWithSelectorShouldRemoveServer(t *testing.T) {
-	_, balanceServer := setupServer()
+	_, apiService := setupServer(&mockClient{})
 	slbConfig := &gen.Config{
 		ListenAddress: localAddress,
 		ListenPort:    defaultPort,
 		Strategy:      gen.SelectorStrategy_SELECTOR_STRATEGY_RANDOM,
 	}
 
-	balanceServer.Configure(context.Background(), slbConfig)
-	balanceServer.Add(context.Background(), &gen.Server{Address: localAddress + serverPort})
-	_, err := balanceServer.Remove(context.Background(), &gen.Server{Address: localAddress + serverPort})
+	apiService.Configure(context.Background(), slbConfig)
+	apiService.Add(context.Background(), &gen.Server{Address: localAddress + serverPort})
+	_, err := apiService.Remove(context.Background(), &gen.Server{Address: localAddress + serverPort})
 	require.NoError(t, err)
 }
